Remove interfaces for converters
This makes RomajiConverter and KanaConverter static helper classes instead
of interfaces. The rationale for this change is WanaKanaSharp was conceived
to be a 1:1 port of WanaKana and introducing such API would deviate it from
the intended goal
